+++ .github/workflows/scripts/npm_release_files/ts_serialize.d.ts
@@ -2,9 +2,6 @@ declare module "@gamebridgeai/ts_serialize" {
   /** A JSON object where each property value is a simple JSON value. */
   export type JSONObject = { [key: string]: JSONValue };
 
-  /** A JSON array where each value is a simple JSON value. */
-  export interface JSONArray extends Array<JSONValue> {}
-
   /** A property value in a JSON object. */
   export type JSONValue =
     | string
@@ -12,7 +9,7 @@ declare module "@gamebridgeai/ts_serialize" {
     | boolean
     | null
     | JSONObject
-    | JSONArray;
+    | JSONValue[];
 
   /** to be implemented by external authors on their models  */
   export interface TransformKey {
@@ -64,24 +61,17 @@ declare module "@gamebridgeai/ts_serialize" {
 
   /** Functions used when hydrating data */
   export type FromJSONStrategy = (value: JSONValue) => any;
-  export type FromJSONStrategyArgument =
-    (FromJSONStrategy | FromJSONStrategy[])[];
 
   /** Functions used when dehydrating data */
   export type ToJSONStrategy = (value: any) => JSONValue;
-  export type ToJSONStrategyArgument = (ToJSONStrategy | ToJSONStrategy[])[];
 
   /** string/symbol property name or options for (de)serializing values */
   export type SerializePropertyArgument =
     | string
     | {
       serializedKey?: string;
-      fromJSONStrategy?:
-        | FromJSONStrategy
-        | FromJSONStrategyArgument;
-      toJSONStrategy?:
-        | ToJSONStrategy
-        | ToJSONStrategyArgument;
+      fromJSONStrategy?: FromJSONStrategy;
+      toJSONStrategy?: ToJSONStrategy;
     };
 
   /** Property wrapper that adds serializable options to the class map
@@ -96,9 +86,10 @@ declare module "@gamebridgeai/ts_serialize" {
    * Converts value from functions provided as parameters
    */
   export function composeStrategy(
-    ...fns:
-      | FromJSONStrategyArgument
-      | ToJSONStrategyArgument
+    ...fns: (
+      | FromJSONStrategy
+      | ToJSONStrategy
+    )[]
   ): FromJSONStrategy | ToJSONStrategy;
 
   /** revive data using `fromJSON` on a subclass type */
@@ -144,7 +135,7 @@ declare module "@gamebridgeai/ts_serialize" {
 
   /**
    * \@PolymorphicSwitch property decorator.
-   * 
+   *
    * Maps the provided initializer function and value or propertyValueTest to the parent class
    */
   export function PolymorphicSwitch(