+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+- removes array as an strategy, use `composeStrategy` instead
+- fmt with deno@1.10.3
+- turn off unstable code coverage
+- remove JSONArray for JSONValue[]
+
 ## [v1.3.1] - 2021-03-14
 
 ## Added